sparklin/JsonParserFuncApp-2/requirements.txt
azure-functions
azure-identity
azure-data-tables
azure-storage-blob
requests